custom widget delete code updated

+++ agile-core-classes/src/com/agilecrm/widgets/CustomWidget.java
@@ -51,13 +51,14 @@ public class CustomWidget extends Widget {
 	
 	public static void deleteCustomWidgetByUserID(String userID, String widgetName) {
 		// Creates Current AgileUser key
-		Key<AgileUser> userKey = new Key<AgileUser>(AgileUser.class, userID);
-
-		CustomWidget csWidget = dao.ofy().query(CustomWidget.class)
+		if(userID != null){
+			Key<AgileUser> userKey = new Key<AgileUser>(AgileUser.class, Long.parseLong(userID));
+			CustomWidget csWidget = dao.ofy().query(CustomWidget.class)
 				.ancestor(userKey).filter("name", widgetName).get();
-		if(csWidget != null){
-			csWidget.delete();
-		}		
+			if(csWidget != null){
+				csWidget.delete();
+			}		
+		}
 	}
 
 	/**
